making : ray_cating
